LTI course creation, allow multiple instructors.
The bridge system was originally designed with the assumption that there
will beone instructor per course. This was modified to support multiple
instructors.

The course information hash's profid entry is now a comma delimited list
of instructor ids. This is passed directly to Webwork's addcourse script
as part of the instructor ids parameter.