LTI launch url fix, LTI prof sync.
The LTI launch url, used for the OAuth signature calculation, was
incorrectly appending a / to the end of urls. This was discovered and
fixed on prod, officially adding it to repo now. The launch url
construction will check the url to see if the last character is /, if it
isn't /, then it adds a / to the end.

The Bridge student import/export structure wasn't capable of syncing
profs. If a new prof was introduced to the course after the course has
been created, the new prof will be added as a student instead of a prof.
It turns out that with the multip prof support for course creation
allowed a pretty easy update to add in prof syncing. Note that profs are
still being added under the P (professor) status and thus are exempt
from being dropped.

Profs can be set the C (enroled) status to remove this exemption.