Reduce number of queries during LTI update.
LTI update is slow. This tries to reduce the number of queries
performed during update to speed it up. Turns out that it was
performing at least one query for every single user_id in the
course to get the User object. Fix was to change the %users
variable from just being a list of IDs mapped to 1 to a list of
IDs mapped to their respective User object.

This consolidates all the individual queries for Users into one
mass query for all the Users in the course.