CourseUpdate, userinfo update, new roles & status.
New role mapping from LTI to Webwork:
    ContentBuilder -> Instructor
	TeachingAssistant -> TA

Note that this only applies to CourseUpdate so far. CourseCreate
will be done later.

We will now compare the user information stored by Webwork and the
info given during the import. If there is a difference, Webwork is
updated with the latest info. First/last name, email, student id,
status, and permission are the only fields to be updated.

New statuses:
EnrolledNoDrop - Students who should not be synced
TeachingNoDrop - Teaching staff who should not be synced

Only the Teaching and Enrolled statuses are synced. So if
there are students or instructors/ta that should not be dropped
during enrolment sync, assign them one of the NoDrop statuses.