CourseUpdate, only update student emails.
There is an email instructor function in webwork. Some instructors
don't want students emailing them from webwork. The workaround for
this is that we just remove their email from webwork. The new
update email function broke that though. So this patch makes it
so that only student emails are updated.

Also make sure not to override emails in webwork when the imported
email is blank.