Autoupdate for LTI.
Similar to autoupdate for Vista. The autoupdate.pl script can now read
LTI credentials from a loginlist file and call the updateclass_lti.pl
script with appropriate parameters to update the course using LTI.

The course update is done by faking a LTI launch request.

The LTI loginlist file is tab delimited and must have the following
fields:

User ID - The user id of an instructor in course being updated.
Course Name - The name of the course in Webwork.
Course ID - The ID of the course in the site we're pulling updates from.
Course URL - The LTI membership request URL.
Key - The consumer key to the LTI secret.
Course Original Name - Name of the course sent through LTI before sanitization.
Time - Timestamp of when this entry was made.

Only the first 5 fields are used as parameters for updateclass_lti.pl.
The last 2 fields are required to be present, but serves only an
informational purpose.