LTI enable writing credentials to loginlist_lti.
When a course is created, we store the credentials used to create
the course in order to perform auto-update for that course later.

Note that I've abstracted updateLoginList out from VistaBridge
and into the Bridge superclass.

updateLoginList takes 2 params, the login list file to write to,
and a pointer to an array of strings which will be combined into
a tab delimited list. In addition to the fields in the array,
a timestamp is written out too.