Autoupdate scripts can check LTI update status.
LTI auto-updates are regular HTTP requests and returns the Webwork
page unless the server is unavailable. We can't tell whether the
actual LTI roster request made by Webwork was successful from the
returned Webwork page without tampering with the single sign on.

This means that we have no way of telling if the auto-updates
failed on any classes. This patch adds a new script,
checkclass_lti.pl, that will perform a LTI roster request using
the information giving in the loginlist_lti and either die if
it was unsuccessful or return quietly if everything works fine.

Autoupdate was changed so that if there is any parameters given,
it'll check the LTI roster requests actually work instead of faking
an LTI launch request.