Show main menu on login and logout page.
The main menu is on the left column that displays a Course link. It used
to show before Webwork got LTI SSO implemented. I think it was removed
not because it conflicts with LTI but that users might be confused at
being able to go to other courses when they got transferred and logged
in using LTI.

This commit restores the main menu since our users aren't always being
logged in by LTI and some got used to having that menu around.

Note, the main menu was hidden by overriding the menu generation code in
the parent class (the links() method).