Enable LTI grade sync.
Note that grades for gateway quizzes are NOT yet synced.

With this commit, we're going to assume that if you have membership
services available as indicated by the LTI launch request, that you
want course enrolment to be updated. As a result, the
$bridge{update_class_on_login} conf variable is no longer required,
neither is the ubc_auto_update parameter in the autoupdater.

Also, if you have outcome services enabled, we will submit Webwork grades
back to the LMS. Note that the current implementation requires you to
have membership services available in order to submit grades.

Note that membership and outcome services are both referring to the
LTI extension services. Formal LTI doesn't support membership yet,
and the outcome service is still primitive. Note that grades submitted
by the outcome service are all in 'ratio' form, e.g. 15/20

Submitting grades require membership services because we need the sourcedid
of the student whose grades are being transferred. The LTI envisioned use
case is that the instructor creates an LTI link that each student will
click on to submit 1 grade back to the LMS. We don't want the instructors
OR students to have to do anything, so our implementation is a bit modified.

The assignment name is concatenated to the end of the sourcedid to be used
for populating whatever grade book the LMS implements,
e.g.: ':_365_1::webworkdev:101:assignment1'

Where ':assignment1' is the name of the assignment. Note that I had to
modify the BBL BasicLTI Tools plugin to support this format.