Refix crash when LTI params contain UTF-8 chars.
I think an update to Net::OAuth we're using broke this again?
My previous fix was no longer working for some reason.

The new fix was recommended by Net::OAuth, which is to use the Encode
library to decode the UTF-8 string to some representation that Perl
likes.

Also fixed error not showing up when OAuth signature checking fails
due to UTF-8 characters.