LTI login ability to go directly to a hmwk set.
Support the parameter "custom_homework_set" which will bring the user
directly to a homework set. custom_homework_set needs to be given
the name of the assignment. E.g.: if the assignment is named "a1",
we'd have:

custom_homework_set=a1

Note that the "custom_" prefix is added automatically by the Basic LTI
tools Building Block for custom parameters.