Sync all grades instead of just non-zero ones.
For each homework, we get 3 values:
'homework_set' => {
	'totalRight' => 0,
	'status' => 0,
	'total' => 3
}

I thought that the 'status' value indicates whether a user has attempted
an assignment or not, so only scores with a non-zero status were sent.

Turns out that even if a student has attempted an assignment, their
'status' might still be set to zero. I think it might have something
to do with the fact that essayAnswers needs to be graded by the instructor
separately.

I've not made it so that grades are sent, regardless of whether or not
a student has attempted the assignment. This means that users will see
a 0 grade for assignments they haven't done yet.