Don't respond to normal LTI outcomes.
Since we've implemented a custom version of LTI grade syncing, we get
unexpected behaviour when trying to send our custom grades when the receiver
is expecting a normal LTI request.

In BBL, this interacted with the changes in the BasicLTI building block to
result in many grade columns being created each time we do an LTI sync.

So now a special parameter is required to enable LTI sync, namely
"custom_gradesync" must be set along with the LTI Outcomes enabled or
Webwork won't try to sync grades.