LTI bridge, get course name from context_title
context_title is the course name in Connect. It is in the general format of
Term-Course-Section-Title-Instructors. The reason we want to switch to this
instead of the course id is that the course id only shows 1 course even for
cross listed courses. Students were confused when they couldn't find their
course in the course list.

The course name in Connect lists all the courses if there are more than one,
e.g.: Term-Course1-Section1-Course2-Section2-Title-Instructors.

We parse the CourseN-SectionN accordingly to make a Webwork course id
that looks like Course1-Section1_Course2-Section2_Term.

Some Examples:
Connect:
2012W2-MATH101-201-Integral Calculus with Applications to Physical Sciences and Engineering-Instructors
Webwork:
MATH101-201_2012W2
Connect:
2012W1-2-AAAA000-000-BBBB111-ALL sections-CCCC222-222, 200-First course: sometimes'dsdfas-Some, Instructor-Moar, Instructors
Webwork:
AAAA000-000_BBBB111-ALL_CCCC222-222-200_

Note that to make sure Webwork course creation succeeds, we have to truncate
the course names at 40 characters.

This commit is NOT compatible with the Vista bridge. The Vista bridge section
will be removed as it's no longer used anymore.