Convert course name convenience script.
This script basically uses the same code path that is used during LTI import to
convert Connect course names into Webwork compatible course names. It's a
convenience script that tells us what the resulting name would be without
having to run the conversion. To use it, just call the script with the Connect
course name as a parameter.