Add ability to redirect to quizzes on LTI launch.
The code is pretty much the same as redirecting to homework sets, just that the
url to redirect to has /quiz_mode before the quiz name.

The parameter is custom_quiz_set, the custom_ part is automatically added by
the basic lti building block, so we just need to put quiz_set into the custom
parameters field.