Option to create student on LTI Launch.
If a student tries to login by LTI from Blackboard to Webwork, but doesn't yet
have an account in Webwork, the login fails. What this commit does is that it
checks for the 'custom_create_student' parameter, which, if set to 'true', will
let Webwork create the student and let the login go through.

The student is created based on their LTI parameters and also assigned all
available assignments. A new sub in LTIParser, parseLaunchUser(), parses user
information from the LTI parameters. The student creation and homework
assignment is done by the CourseUpdater's addUser().

To enable this in the BasicLTI building block, add this to the custom params:
create_student=true

The assumption here is that since the student already has access to the same
course in Blackboard, they'll eventually be synced to Webwork anyways. Note
that this will only work with new students. Returning students who have dropped
the course will not be reactivated until the regular sync finishes.