Import: log class count, fix drop msg role.
Add studentupdates.log to give us a summary count of the number of people
currently enroled in a course, currently dropped, and the number of people we
just got from LTI import.

The drop messages doesn't distinguish between roles while the rejoin does,
fixed so that when a teaching staff is dropped from a course, it'll properly
log it as a teaching staff dropping a course instead of a student.

Note that new users are still all logged as students, probably need to fix that
too.