Add Webservice call to retrieve detailed grades.
LTI grade sync can only sync the total grade for an assignment. It'd be
nice if we could get more detailed grade information at an individual
problem level (e.g.: # of attempts each problem, which problems were
right, which were wrong). Since all this information is already stored
in the database (and can be viewed by instructors in the web interface),
I just modified the existing grade retrival functions for Webservice.

Unfortunately, this introduced code duplication, since I've already
copied the grade retrival code for LTI. What should be done is to change
the LTI grade retrival to rely on the Webservice call if possible.

What would be really nice is if Webwork itself is reworked so that grade
retrival is in one central location. As it is, it appears that grade
retrival happens in several locations and they all share pretty much the
same code. I'm guessing this is happening eventually as there appears to
be active effort at completing the Webservice api. At which time, I'll
come back to revise this implementation.