LTI update check students are in the right course.
Basically, the sync is somehow pulling in students from one course into
another for some reason. So the existing students who should be in the
course gets dropped and then then the wrong students from the wrong
course got added.

The LTI parser will now check that the students are in the right course
and if they're not, will error now and abort the update.

Also added additional tracking to help figure out what's going on. We'll
know the course that the update requested, whether it was the
autoupdater or an instructor who triggered the update, and if the update
aborted, what course did the student come from.