Update conf templates. Restore math4 to upstream.
New additions conf.dist files was reflected to our UBC specific dist
files.

There is a new feature which will use git to check for updates to
Webwork. I've disabled this for stability reasons. The release note
states that there's a risk where "If the wait is too long the system can
be disabled".

The math4 template has had some changes. I had previously hacked in some
fixes to the banner in order to make it not create a horizontal scroll
bar. This fix didn't update well, so I've restored the upstream version
of the math4 theme. I've also hacked in a dirty fix instead of the more
involved fix I had previously to prevent the creation of the horizontal
scroll bar.