Add postmethod to problem submission
Extremely odd situation where a student was submitting answers with `get` instead of `post` for a few (but not all) courses. This occurred on multiple devices for the student but other logging in as the student would not run into this issue. A general workaround is add a redundancy formmethod="post" to the submit button on the page.