FIX delayedjob worker fails to complete long jobs
It looks like on long running jobs, the mariadb database handle used by
The Schwartz job queue gets disconnected due to inactivity. The
executing job doesn't use the same database handle, it uses one created
by Webwork's course environment, which doesn't get disconnected because
it gets used a lot. When the job completes, Schwartz goes to try to
remove the job from the queue using the disconnected database handle and
errors. This leaves the job in the queue for another worker to pick up,
even though it was actually completed successfully.