ADD lti score canvas-specific submitted_at param
Scores submitted via LTI AGS to Canvas have their submission time set to
the time that Canvas received the score. It does not go by the
"timestamp" parameter sent in the score for submission time. This meant
that when we have grade sync delays and the score ends up being received
by Canvas *after* the assignment due date, Canvas wrongly marks it as
late.

Turns out there's a Canvas specific extension to LTI that lets us set
the submission time, we just need to add to the score json:
  {
    "https://canvas.instructure.com/lti/submission":
    {
      "submitted_at": "2021-09-21T18:20:47.000-0700"
    }
  }