FIX random caliper jobs can't be executed
Rather than a fix, this is more of a workaround. The issue seems to be
with TheSchwartz library's use of the Storable library to serialize and
deserialize perl hash representations into binary blobs. So the
workaround just avoids giving TheSchwartz a perl hash to serialize. We
put all our data into a JSON string and just give TheSchwartz that
string. Hopefully this fixes it, it worked for a few jobs that I
manually converted.

Things that doesn't seem to be the issue:

The serialized representation might have elements in different orders,
since the hash representation itself in perl doesn't guarantee order.
But this doesn't seem to be an issue, since I can find both working and
broken jobs stored in the same order.

The caliper events can include the entire pg file as part of the data. I
thought this might be an issue because TheSchwartz uses eval { } as part
of the deserialization. However, I found working jobs with pg files
included and also found some broken jobs that doesn't include pg files.

I thought there might be some UTF-8 handling trap causing issues, but
then I found broken jobs with contents that only use ASCII.

There might be a collision issue with two workers picking up the same
job. However, when I tried to duplicate two workers picking up the same
job, it always resolved properly with one worker actually completing the
job and the other worker dropping it. Furthermore, a known broken job
does not load properly, even if you isolate it and ensure that only 1
worker picks it up.

It doesn't seem to be an issue with the database. By modifying
TheSchwartz library, I was able to check that the data returned from the
database was correct. It is only after TheSchwartz called Storable to
deserialize the stored data that it disappears. Turns out Storage
returns undef on what it calls "normal" and "internal" errors. And there
seems to be no way to get it to cough up error messages.

Manually inspecting the mysqldump of working and broken jobs, I can find
very similar ones that only minorly differ (same course, same event,
different user), yet one is working but the other not.

Maybe there is something about binary handling in perl or the binary
representation stored in the database that's the key to all this.