EDIT lti ags score sync timestamp set to cur time
The timestamp used to be set to the student's assignment submit time.
This caused a situation where if the instructor changes the Canvas
assignment configuration, Canvas will reject our grade sync request
since the LTI Score happened before the Result's updated_at timestamp.

To avoid this scenario, we'll set the timestamp to the time that the
sync is sent. Assignment submit time is sent using the Canvas custom
param, so shouldn't be affected.

Unfortunately, this does count as a "new" attempt on the assignment, so
the instructor might need to increment the allowed attempts.

This behavior seems to contradict the fact that the timestamp is
supposed to be "Date and time when the score was modified in the tool"
according to Canvas documentation. It seems like there was 2 different
ideas of what role the timestamp should play in the Canvas LTI
implementation. Not much we can do about it.