ADD new permission level "designer"
Designers were previously mapped to "professor" and there were concerns
that this is too much access. So it was suggested that we have a role
that can create content but doesn't have grading power.

I initially thought this would be at a level below professor but above
TA. However, it looks like over the years, we've given TAs a lot of
professor permissions as some profs prefer TAs to handle Webwork. So we
had to put designer at a lower level than TA instead.

The LTI role ContentDeveloper is now mapped to this new designer access
level.