FIX ContentGenerator.pm incomplete merge
Some of the merge blocks in ContentGenerator.pm weren't dealt with.
Looks like the only big change was the addition of icons to the Log In
and Log Out links. Since we customized the Log In/Out links to show the
user's name instead of their user id, it showed up as a conflict.

Also added in the result of merging localOverrides.conf.dist with
ansible's localOverrides.conf template to make the two match up more.