FIX Dockerfile-prod missing some dependencies
There were changes in apt packages and also the addition of npm package
management from Dockerfile that needs to be in Dockerfile-prod.

apt packages removed:
    libdancer-perl
    libdancer-plugin-database-perl
    libmoox-options-perl

apt packages added:
    dvisvgm
    libarchive-zip-perl
    libcgi-pm-perl
    libjson-maybexs-perl
    libcpanel-json-xs-perl
    patch
    pdf2svg
    iputils-ping
    imagemagick
    npm

Looks like upstream webwork has moved to npm for frontend packages. So
we needed to add an npm install step to Dockerfile-prod. In the upstream
Dockerfile, this was done as part of phase 5's config changes. However,
I've split it out into it's own block as this is a big step and should
be executed by itself to take advantage of caching.