FIX assign to all exception, proctor pw styling
When clicking "Assign to All Current Users", you might see an exception
about status_abbrev_has_behavior()'s first parameter cannot be null.
This is caused by proctored quizzes creating a proctor user for the
proctor login and setting this proctor user's status to null. The
proctor's user status is null because we deleted the proctor status in
localOverrides.conf. I've restored the proctor status, setting the
status abbreviation to "G" since we're already using "P" for profs. Note
this special proctor user does not show up in the classlist, only shows
up in the database.

While creating a proctored quiz, I noticed the proctor password field
wasn't bootstrap styled cause there wasn't a type attribute on the
field. I've set the field type to "text" to fix this. I didn't want to
set the field type to "password" in case it changes the behaviour too
much. I've also added a class "input-medium" so that the field is a bit
wider, this overrides the previous width set by attribute 'size="10"'.