FIX GatewayQuiz autosave not working
The autosave basically pretends to be a preview submit. It looks like
the preview submit now expects the value of the "Preview Test" button to
be submitted. The value of this button is treated as a marker for a
preview submit. However, the value of the preview button is only
transmitted if the button was clicked to trigger the submit. This means
when we serialize the form data, the button values are not included.

I've added a workaround to just append the preview button value to the
the form data. An interesting discovery is that the value doesn't
matter, as GatewayQuiz.pm only checks to see that it's not empty. So we
can set the value to "autosave" to distinguish between a user triggered
submit and an autosave.