FIX designer cannot edit pg files
This is due to our previous change of removing 'modify_student_data'
permission from designer. Turns out the PGProblemEditor checks for the
'modify_student_data' permission for many operations. I'm not sure why
it's doing this, so have asked the Webwork github just in case there's
some special reason behind this.

For now though, to fix this quickly, I've changed it so that instead of
checking for 'modify_student_data', it checks for 'modify_problem_sets'.
This makes more sense to me on the surface, since editing problems have
nothing to do with student data and if you can edit problem sets, you
should be able to edit the individual problems too.