FIX lti classlist&grade sync runs in infinite loop
When executing the classlist & grade sync cron scripts located in
lib/LTIAdvantage/Cron/, the docker container doesn't terminate after the
scripts terminate, it instead just keeps rerunning those scripts. This
keeps adding a ton of classlist & grade sync jobs into the job queue,
overwhelming the job worker.

Turns out docker-entrypoint.sh was changed to run in an infinite loop in
one of the upstream webwork commits. Looks like they were trying to
restart Apache without having to restart the container. I'm guessing
this has something to do with using the docker-compose as a development
instance, where they want to restart Apache so it'll pick up Perl code
changes. Maybe they had some issue with Apache::Reload and had to use
this approach?

In any case, this isn't a valuable change for us to retain, since we're
using the docker image in actual deployment. So the easiest solution was
to go back to the old code before this change.