FIX Hmwk Sets Editor slow (~30s) loading time
This is from "Import" tab where it scans the course templates directory
for set def files. It does this in such a way where it wants to follow
symlinks, since the OPL and other libraries are symlinked in course
templates and there might be set defs in there that the user want. It's
this follow symlink option that seems to be causing the slowdown.
Removing the option entirely restores the page load time back to normal.

The "follow_fast" option in the original line slows down our prod
servers a lot. Apparently, follow_fast "may report some files more than
once" according to documentation. This may be due to some interaction
with our NFS mount courses storage, but what happens is it starts
repeating the course templates dir and then seems to get stuck for ~30s.
So pages that call getDefList gets slowed down a lot.

An alternative to the "follow_fast" option is "follow". Using "follow"
=> 1 appeared to require "follow_skip" => 2 or we'll see an exception.
"follow" and "follow_fast" seems to be equivalently slow, so looks like
we want to disable it entirely.

This means that we won't be looking into the OPL or ubcLibrary for
set def files, but we don't seem to have picked any up before this
anyways.

The "useOPLdefFiles" config option is meant to control whether we look
into the OPL for set defs, but it does this by skipping the library
directory. This option now does nothing since we're now skipping
symlinks.