FIX worker crash, hide chg pw, Grades.pm crash
Worker was crashing due to the WWSafe module being moved to the PG
library. CourseEnvironment imports WWSafe, which is used by the worker
script run_all_jobs.pl. Need to import the PG library in run_all_jobs.pl
so that it can find WWSafe.

We were hiding the change password functionality in Webwork by just
commenting out the code. The Bootstrap 5 upgrade resulted in a lot of
conflicts due to this. Not sure if this is new, but it looks like the
change_password permission can be used to hide this instead. So I've set
change_password to only admins and above. Since only admins should be
using the local Webwork login. Everyone else should be using CWL or LTI.

Grades.pm just had a variable renamed.