FIX StudentProgress number of attempts column
The grade_set() function that used to be StudentProgress.pm was
customized to return number of attempts. This grade_set() turned out to
have many duplicates and was finally abstracted out into Util.pm in
Webwork 2.17.

I've re-customized the one in Util.pm to also return number of attempts.
But I'm not sure if this is the best move. I also tried adding the
number of attempts on as a separate function, but unfortunately, it
looks like it would involve undoing the point of the original
deduplication effort. Though this is probably due not being familiar
enough with webwork.