EDIT style & bug fix for quiz,progress,problemset
ProblemSetList: The instruction "Please select action to be performed"
should not be in a green alert box. The alert boxes should be used to
point attention to the results of interactions and no interaction has
occurred here.

GatewayQuiz: handle_input_colors function was deleted, so we don't need
to call it.

StudentProgress: $recordAsOther and $recordVersionsAsOther are both
redeclarations and can be deleted. Styling updated to bootstrap 5. I
also tried to make it look more separate from the table above.