EDIT envVar api url for opl table dumps/json files
bin/download-OPL-metadata-release.pl has a hardcoded API URL for
downloading the table dumps/json file release archvies generated by
github actions. This is hardcoded to the official repo, but we want to
use the ubc fork. I've set it to default to the ubc opl fork, but it
should also be overrideable as an environment variable, just set
OPL_REPO_RELEASE_API_URL

Not deployment related, since it's only used for dev, but I've also
switched over Dockerfile-prod to the ubc opl fork. I've also omitted
deleting the .git/ directory from the opl, as
download-OPL-metadata-release.pl expects the opl directory to be a git
repo.

Also not deployment related, I've limited npm to --maxsockets 2. This
isn't an issue on campus, but I found that my home internet really
didn't like opening so many sockets.