FIX pod being killed prematurely on startup
The docker-entrypoint.sh became heavier this version and takes a longer
time to start up. I think this means that the pod was failing to come
online fast enough to pass the healthcheck, which only gives the pod 30
seconds for startup. I've added an initialization time of 5 minutes,
where failures of the healthcheck wouldn't count against the pod, this
should give the pod enough time to run all the init.