FIX webwork stg worker crashloop missing site.conf
Webwork staging worker was in a crashloop due to missing site.conf.
These conf files are supposed to be created by the docker entrypoint
script. Turns out the conf file creation was behind a check to see if
we're executing apache. Since the worker doesn't execute apache, it
doesn't create the conf file.

The merge from upstream undid a customization we had previously where
the conf file creation was executed without going through the apache cmd
check. This commit restores that customization. Note that I've disable
the localOverrides.conf section as our kubernetes deployment uses a
configmap for it.