FIX mathview button floating above problem nav bar
When looking at a regular homework set (student view), the MathView
button floats on top of the problem nav bar when an answer box scrolls
below it. The problem nav bar is itself a floating element attached to
the top of the window. The MathView button appears above this nav
because it has a z-index of 2 from stock Bootstrap. Resetting the
z-index to auto fixes this.

Also added a little bit of spacing around the MathView answer boxes, as
I notice they were squishing up against the equal sign and other answer
boxes above and below it.