FIX "grade saved drafts" doesn't save submit time
The "Grade saved drafts" button on the "Student Progress" page for a
quiz is supposed to batch grade quiz attempts where the student entered
answers but for some reason was unable to submit themselves.

This grade saved drafts function does not give the quiz attempt a submit
time.  The submit time is recorded in the set_user table's
'version_last_attempt_time' column. When this column is empty, students
will see a "No Submission" in their homework list for the quiz. To
reduce confusion, the grade saved drafts function now sets
'version_last_attempt_time' to the current time so that students
actually see that the attempt was submitted.

Just in case it might be useful, I've also added an identifier for
quizzes that were graded via the "Grade saved drafts" button. A string
"(Graded via saved draft)" is appended to the 'description' column for
the student's quiz attempt.

Note that the "Test Time" column on the "Student Progress" page defaults
to the quiz's time limit if the time taken was greater than the time
limit. For grade saved drafts, the time taken should always be greater
than the time limit, so the "Test Time" column should always be the time
limit.