FIX lti launch fails due to missing state cookie
This happens when the LMS is on a different domain and is due to
SameSite Lax preventing the state cookie from being presented by the
browser.

A short term solution would've been to just switch the LTI Launch cookie
to SameSite None, but that's not recommended for security.

The alternative is to go cookie-less. As far as I can figure, as long as
we have something that serves as a CSRF token, it should be ok security
wise.

In order to serve as a CSRF token, we need to generate cryptographically
secure random strings. We're using the new 'Bytes::Random::Secure::Tiny'
library for this purpose.

We were previously using 2 UUID strings for state & nonce, I don't want
to exceed this string length for the CSRF token. We're passing auth
redirect params as GET params and this makes for very long URLs that
might hit URL length limits. This means 32 character + 32 character = 64
character max for our CSRF token.

We'll generate a 64 character string and split it in half, one half for
state and one half for nonce. Note that while we'll be passing state and
nonce as separate params, we're storing the entire 64 character string
as one entry in the nonce table.