EDIT error table needs index on jobid, error_time
Error tabled filled up to 300k entries and started lagging the db
replication. The db slave couldn't keep up for some reason, taking
several seconds to process deletes to the error table.

None of the columns in the error table is indexed, which is our guess at
why it's having speed issues. It seems jobid isn't unique, as multiple
failures can result in multiple messages. Such duplicates seems to be
deleted quickly but this still means that jobid itself cannot be unique.

So this should be a compound index on jobid and error_time. We actually
also end up with an index on just error_time due to the weird way that
Webwork handles indexing, not worth the effort of trying to prevent
that, so just noting it here.

error_time is a unix timestamp, so we only have seconds level
granularity. So it's possible that an immediate retry of the same job
results in two identical (jobid, error_time) being inserted at the same
second. This is avoided by making sure DelayedJob::Worker::retry_delay()
returns at least 1, so it'll wait at least 1 second before trying a
failed job.

While Webwork's db upgrader can automatically add columns, it seems that
it doesn't handle adding indexes. So if there's an existing error table,
it'll need to be upgraded manually. Either manually add the index or
just drop the error table and login to the Webwork admin course to
recreate it.