FIX caliper job error when some utf8 char are used
The problem text used ★★☆☆ which errored out with HTTP::Message
complaining that the message contains bytes. Initially thought it was
the database mangling the encoding, since the db defaults were latin1.
Converting the database to utf8mb4 didn't solve the issue though. Turns
out LWP::UserAgent / HTTP::Request downgrades to latin1 encoding by
default. To force it to use utf-8, we have to explicitly encode to UTF-8
and then tell it to use UTF-8 charset.

Note that in Perl, 'utf8' and 'utf-8' are two different implementations
and are not interchangeable. 'utf8' without the dash is an older
implementation that is no longer compliant with current utf8 rules.
'utf-8' with the dash is the newer, more current utf8 implementation.