EDIT lti ags 404 resource links marked as invalid
Resource links are invalid when the corresponding Canvas assignment is
deleted. This is actually an existing mechanism, but in NRPS. NRPS
returns an "invalid rlid parameter" error for deleted assignments. I'm
not sure if this is another consequence of the Canvas feature flag
ags_improved_course_concluded_response_codes, but AGS now seems to tell
us about deleted assignments with 404 error that has empty content. The
NRPS error doesn't seem to happen anymore.

Still keeping the NRPS implementation for Canvas instances that doesn't
have the ags_improved_course_concluded_response_codes feature flag
enabled.

AGS requests will not be generated for resource links where is_valid is
false. So we just set the resource link to be invalid when we get a 404
with an empty content.