RENAME LTIAdvantage to LTI1p3 (conflict likely)
A separate commit for the rename where we have the highest probability
of conflicting with the upstream during the merge.

The LTIAdvantage namespace is also used by upstream Webwork. To prevent
stepping on their implementation and to distinguish between our
implementation, I want to move to the LTI1p3 namespace.

The most likely conflict is in lib/WeBWorK/Authen/

I'm hoping that doing the move as a separate commit allows us to keep
history on the LTIAdvantage.pm file when we merge with upstream.