FIX compile fail: validate_json(), Cookie,mod_perl
All compilation failures are in lib/WeBWorK/Authen/LTI1p3.pm

JSON::Validator removed the validate_json() call. New way is to create a
validator, pass it a schema, then call the validate method.

WeBWorK::Cookie was removed, we were importing it but wasn't really
using is, so we can just remove the import.

mod_perl is also no longer used and was removed.