EDIT lti1p3 Entrypoint::Login to mojo controller
Mojolicious has several options you can enable to automatically set perl
feature flags. E.g.: use strict and use warnings is the '-strict' part
of 'use Mojo::Base'. '-signatures' enables the subroutine signatures
feature: 'sub accept ($self)' declares that it expects the parameter
$self in the first param, making it more similar to other languages. The
'-async_await' flag enables declaring subroutines to be async.

Important, route handlers in controllers needs to be async. Otherwise,
exceptions gets silently lost. This has something to do with the router
expecting to handle promises. When the controller is handed a non-async
handler and it throws an exception, it just treats it as a promise that
never gets fulfilled.

It doesn't seem like Webwork is really taking advantage of the async
aspect as database operations still seem to be synchornous. So I didn't
make any special changes to adapt for async.

Found out that returning proper http error codes require using
$c->reply->exception('msg')->rendered(error code) so converted all the
errors to that.

Realized that the redirect_uri param was being lazy and using
target_link_uri cause the old implementation uses the same url for all
stages of the launch. Now that we have specific urls, we have to
generate a different url for it.